#13, #14, #24, and more fixes
Added more lights to HQ even though I really don't see a big difference.
Added feature to be able to replace guards as well as stationary weapons
for HQ and FOBs, but not replace the building itself, I thought it would
be too OP. Added autosave in singleplayer when accepting missions.
Finally, fixed an array problem in manual zones.