New Heli Taxi System
Fixed pretty much everything from the pull request I closed except for
the pilot's dialogue (the FOB name is different every time) and there
are some duplicated markers until the taxi returns to base. Other than
that it works fine. You must call the taxi for the first time while
you're at the base or else it'll always return to the location of its
first request. Let me know if there are any big issues that it screws up
and I'll do my best to fix it.