Sell Resources and Taxi Fixes
Very basic "Sell Resource" script. It only checks the condition of the
engine as of right now. Wheels are finicky but I'll figure something out
eventually. I'll also eventually add rear rotor for helis and etc for
certain vehicles. As for the taxi, you no longer have to call it from
the base first. It'll fly straight there on its first run.