Added EventHandler for negative playerRating
When player rating goes below -2000, the player is deemed 'Traitor' and is
treated as an enemy by ALL factions and KOS.

This EH checks for negative rating (currently below zero) and resets back
to zero.
If the future rating value is negative, the rating change is set to zero
else continues normally.

Referencing: DUWS-R-Team/DUWS-R#151