v0.7.0
There may be some changes I've forgot to list
##Additions
- Functions Library (Replaced all cases of ExecVM)
- Simple replacement of units using DUWS_CONFIG.SQF
- AIS Revive
- BLUFOR Commander now has static animations

##Changed
- Purchase menus have a length equal to the array found in DUWS_CONFIG.SQF (Purchasing vehicles, units, groups)
- Imported mission file to 3DEN
- Cleaned up mission root to some degree

##Deprecated

##Removed
- Old DUWS Revive system

##Security
As there are some heavy changes present, there are bound to be some bugs and issues. Some things are probably getting called when they should be spawned, and vice versa (If it could be called, it is called else it is spawned). 

This version can likely be seen as an Beta build.

##Todo
[] Thoroughly test this
[] Test on dedicated server
[] Thoroughly test unit/group configs
[] Cleanup codebase