spellcheck, FOB requesting and SITREP on dedicated servers
On a dedicated server, SITREP and Requesting a FOB never appears. This looks to be because BIS_fnc_addCommMenuItem is local, so needs to be called on the client, and i cannot for the life of me see if it already is. So i've just added it to the bottom of fn_clientInit.sqf.